Read db/table descriptors from system config on common ops.
For `SELECT`, `INSERT`, `UPDATE`, and `DELETE` only.

Planners get a copy of the gossiped system config at execution time.
DB and Table descriptors are read from the cache.

If something is not found (or is corrupt) in the cache, fall back to the
standard lookup. However, if a database descriptor is not found in the
cache, do not attempt to lookup the table descriptor in the cache.

Bank example benchmark comparison:
```
benchmark              old ns/op     new ns/op     delta
BenchmarkBank-8        2546883       586224        -76.98%
```